Restructed project and added draft for instance ui management
Restructured project in a way that namespaces now match the paths and also
splitted classes for operator management mostly into own files.
Added concept of InstanceUi class which holds the ui relevant information
for an instance. This is kept in a global registry and can be accessed via
the id of the symbol that contains the instance and via the instance id
itself. See comments in InstanceUiRegistry.