Fixed problem when setting a new default value across several instances
Handling default values is a bit tricky with ImGui as we want to show the default
value when this is set, but we never want to default value to be modified. But as
editing is already done when the return value of the ImGui edit control tells us
that editing has happened this here is a simple way to ensure that the default value
is always correct but edition is only happening on the input value.