Inputs and outputs of instances are now automatically extracted
For that a generic typed Instance<T> type was introduced from which the
operators now derive. Doing this enables to do the extraction in the base
class as the type of the derived class is then known.
Extraction of the inputs is done by checking for IInputSlot and for the
outputs by looking for the OperatorAttribute with the output type set. This
is for testing and should be unified later when it's clear which way to go.