Fixed input/output abstraction and made composition op input/outputs connectable
There are now new IConnectableSource and IConnectableTarget interfaces to
mark slots which can be connected. Also the default Slot<T> behavior is now
already supporting connecting it. The default `UpdateAction` is now just
checking is there's a connection. This is the usecase for composition ops.
In case of output slots within a code op these are setting their own
`UpdateAction` anyway. So the new solution supports both ways of usage
without a large change.