Cleaned up mock model that now the symbols are generated automatically
Now the core assembly is searched for types derving from Instance and not
being a generic type (for now in order to prevent Instance<T> to be found).
For all these type a symbol and all the corresponding ui registry entries
are created.
For now the 'Project' op is then manually placed in the 'Dashboard' op.