Replaced the global Registries for InputUis, OutputUis and ChildUis with SymbolUi
The global ui registries for inputs, outputs and children have been removed
and also their storage files. Not the symbol uis are saved and loaded
like the core symbols. At the same place with the same name but with the
'.t3ui' extension. All code for the former registries has been removed.

Todo: The previous existing automatic generation of inputs/outputs/childs
has been moved to SymbolUi but not enabled so far. This will be done in
a next step.