Fixed exception when creating connections
Problem appeared when a connection was creating to an input of an op
which type was already used in the composition. Reason was a missing
check for the right instance id of the target op. Without this there
were then simply more than one target inputs with the same id so that
SingleOrDefault failed.