Fixed setting input values again with redo
This happend when the previous state was the default value. In that case
the 'redo'd value was set but as the input was still set to 'default value'
is was not shown in parameter view.

Todo: default value is currently not really handled in commands and
this must be added.