Fixed crash when input parameter view tried to show connected values
Problem was that it could happen that the selected op was never evaluated
so far. In that case the Value property could still be null when reference
types were used (e.g. List<>). Now the input parameter view updates the
input before displying the output (just like the output view does).