Creating now simpler update action with a specific use case for each
Previously the Update() action of InputSlot was handling everything,
if it is connected and if not extacting default of input specific value.
Now this has been split, when an input slot is connected always the
connected value is used, so there's no need to check this. Therefore
when an input slot gets connected now this simply sets up an update action
that is only requesting the connected value. The default update of an
input is now only checking if default value is set or now and then just
returning the corresponding value. It doesn't check for connections anymore
as this is now handled in the connection specific update action.

This enables having specific update functions for specific use case, e.g.
animated input slots will set an animation specific update action.