Fixed showing the 'normal' inputs of a selected op if another shows input parameter
Problem was that when for an input the parameter view of this input is shown
and then another op was selected in the parameter view neither inputs nor the
input parameter edit were shown - only the common properties. Now showing
the input parameter edit is only done if the correct op type is selected.
Otherwise the normal inputs are shown.