Fixed crash when stepping into an operator
Problem was that the composition op was already changed but the code
still iterating over the 'old' children. For now it's just allowing null
value when searching for child. Later on this should be fixed by having
some kind of command list that is executed at the end of a frame when
drawing the current frame is done.