Fixed crash when doing combineAsNewType
Problem was that intputs and outputs were generated that were not valid
one. There was a check missing if the source/target of a connection to
an op that is going to be combined also in the selection of combined
ops. In that case the connection is not used to generate inputs/outputs.