Added first draft of bass audio integration
The native bass.dll must be copied to the T3 build folder. Beside that
the original ClipTime has been made abstract and there are now two
implementations:
- one that behaves like before (UiClipTime)
- one that encapsulated the bass playback (StreamClipTime).

In constructor of GraphWindow (where clip time is created) it's checked
whether the set mp3 is in 'Resources'. If that's the case then the
StreamClipTime is created, otherwise the UiClipTime is created.