Fixed crash after deleting an input
This is NOT fixing the problem itself. It's just preventing the crash and
writing out a warning to the console. If this happens go up to the parent
op and back into the original one.
The error happens when an input of the actual composition op is deleted.
The reason is that the GraphCanvas holds a reference to the composition op
instance and this is not updated when a new instance gets compiled.