Fixed crash11 after adding a new input
Problem was that there were two ops of the same type (e.g. CurlNoise in
PartialScene02) where the output of both were connected to the same
multi input. In that case it could happen that the order of adding the
connections again after the instances have been replaced were wrong. In
that case e.g. a connection with multi input index 2 and 4 were
connected (5 connections at all to the multi input). So the connection
with 4 was the last one. After both were removed it's mandatory that the
one with index 2 gets added first as otherwise the index 4 is out of
range - only 3 connections there when trying to add with index 4.
Now the order is reversed before adding the connections.