Fixed crash 14 - disposing shader resource which is in use
Problem happended when a shader was used from at least 2 ops. In that
case after a change of the shader source the first op was updating the
resource. Which means disposing the previous stat and compiling the new
one and setting this as it's current output. Then the second op was
doing exactly the same and so disposed the shader from the first op.
Now the shader is only compiled and the previous one disposed on the
first update. For the second op nothing is done as the shader is
already up to date.