Fixed evaluation of inputs of compound ops
Problem was that instead of using the dirty state of the input directly
(if it was not connected outside) the routine for the whole parent op
with all its inputs was used which resulted in dirty states for simple
inputs where the value never changed.
Now this state is checked explicitly and the dirty flag is directly
used.