Fixed crash when combining to a new type
Problem was that part of the combined ops were two times the same type
(in that case Value) which input and output was connected. The crash
resulted from the fact that for both instances of Value the input ID is
the same. Internally a mapping was created from the input ID to the
newly created input. But as the ID was the same trying to add the same
input ID a second time as key resulted in an exception.

Now instead of adding these input IDs to a map (which is not possible
cause of the above mentioned problem) the original connection (the one
pointing to the original input) is stored together with the new input
ID. Now when the new connection is created - the one to the new input -
the old connection is used to look up the new target slot ID.

The same was done for outputs.