Fixed id-path generation for depth larger than 2
Problem was that in the end only the root and the searched elements were
added. But with a nesting depth > 2 the elements inbetween were missing.
This has been fixed. Now first it's searched until the searched element
is found. This is added first. When going the search path back upwards
all upper levels are inserted to the front of the list.