Added BeatTime to context
This is synced to BPM and clip time. It keeps on playing even in pause mode (so it cane be used like runTime) but it will be resynced to cliptime once playback starts. This makes it easier to write shaders that are synced to beat and independent of BPM rate.