Update code recompilation scheme after input add/remove
The code is now still recompiled and the symbol is updated with the new
instance type. But the source isn't automatically written to disk. It's
stored in the symbol until a save is done.
This was done in order to keep symbol source and meta files in sync and
also to have a consistent save behavior.