Duplicate as new type now uses roslyn for creating the new source
This prevents possible errors due to the simple search and replace
previously used (e.g. duplicating 'Value' wasn't possible). Now with
roslyn it's explicitly searched for class, ctor and input/output
declarations and modify only the relevant parts.
Beside that the new source is not saved automatically anymore. This is
now also done only on save (like for add/remove inputs so far). This
also means that the new operator type is added to the operators project
on saving the first time after duplicating.