Fixed crash when using variations and operator preview
Problem was that the ImageOutputCanvas was not using the SrvManager but
instead managed a SRV on its own. This lead to the problem that when
changing the preview op a previous SRV was disposed which was still in use
for drawing at another place.
Now the ImageOutputCanvas also uses the SrvManager to prevent such things.