Fixed non working copy and paste of operators
Problem was a regression when removing the instance type from the serialization
of symbols. The container symbol for the copied content uses typeof(object)
as instance type and this wasn't considered anymore. Now the caller of
ReadSymbol can specify that non operator instance types are allowed. In that
case the object type is used when the type couldn't be found in the operator
assembly.