Fixed crash when resizing the window
Problem was that the RenderTarget op was requesting the actual set RTVs
in order to restore them at the end. Requesting them increases the ref
count for the resources, so afterwards a dispose is needed. This was
missing and resulted in pending resources to the swap chain buffer - the
swap chain RTV. The swap chain then throw an exeception when trying to
resize it as there were pending resources.
Now the RTVs are disposed after restoring them (and therefore the ref
count is decreased).