Fixed assert when duplicating as new type
Problem happened when duplicated type had animations. Problem was that
nowadays the animations are already copied in the
CopySymbolChildrenCommand. So in DuplicateAsNewType was again calling
Animator.CopyAllAnimationsTo(...). And this method has an assert that
the target animator must be empty.
Now the additional copying is removed and the animations are only copied
due to the command.