Fixed rendering the wrong texture into canvas texture
Problem was that the output-ui texture and its SRV was cached. But as rendering
of variations can happen with different resolution this led to the usage of
the wrong texture when rendering into the canvas texture. Now the texture
and SRV are taken from the slot directly after evaluating it with the current
variation. This way always the correct texture is used.