The dirty flag trigger is now configurable
The default value for outputs is 'None' but the outputs can overwrite
the default value with an attribute. This default value is used for all
symbol children. Additionaly each symbol child can configure a different
trigger type in composition context. This is done via the RMB menu on an
output.