Reworked the 'always' behavior of the dirty flag
Previously a slot with a dirty flag trigger 'always' was propagating the
dirty state up during invalidation. This meant that an [Execute] below a
[RenderTarget] was always setting the render target to dirty even if
nothing changed below the [Execute] and the render target result was
still the same.
Now outputs with 'always' are still always evaluated if the result is
requested, but these outputs doesn't propagate the dirty flag upwards by
themselve. This only happens if the flag was set to dirty from downwards
the graph.