Fix for incorrect calculation of 'updates since last frame'
Problem was that the previous calculation was bound to the evalution time
of the output window within a frame. There was a correction offset in the
calculation that took into account the usage of the value was happening
one frame later (which equals 100 - the GLOBAL_TICK_DIFF_PER_FRAME). But
when requesting the property within the same frame a wrong value was
calculated that was to large by 100.
This has been fixed by checking if the global tick count is already in
the next frame (diff >= 100) or not. Now only in the first case the correction
is added.