Handling editable input UI reference types is done now in a more generic way
The actual behavior didn't change, for some input types a clone is created
if the input is pointing to the default value (e.g. Curve). But now this
is done the the base InputValueUi class. In order to detect the relevant
types IEditableInputType has been introduced which derived from ICloneable.
Input values of types that implement this interface will be cloned when the
input is set to default.
ToDo: this needs somehow to be extracted to be used at other places too, like
      the in op editing of input values.