Fixed error that symbol id was wrongly generated if no meta files exist
Problem was that the ID is nowadays part of the symbol instance type
namespace. So if the type is existing but no meta files the id from the
namespace must be used, otherwise there will be different IDs and next
time loading will fail.
Now in that case the ID is extracted from the namespace.