Fixed op compilation with missing assembly references
Problem was, that some used asssemblies reference itself assemblies that
are not part of the Operator assembly dependencies. In that case the
compilation failed. For now also all the references of the used assemblies
are added too. This seems to work. Potential problem as this solves only
one level of indirection: the newly referenced assemblies could itself
reference things that were not added so far.