Fix multi evaluation per frame bug
Problem was that the input dirty flag was preventing the re-evaluation of a connected input when called more than once a frame.
Now the evaluation is done always when the input is connected. This should be right as the connected output is defining if the output conaining op should be re-evaluated or if its output is cached.