Fix multi invalidation of default cached inputs
Problem was that previous check to see if an input/output/op was already visited during
invalidation was only set when invalidating that input/output/op. Therefore for trees with
no animation/always-dirty set this check always failed which lead to re-invalidation of
sub-trees. Now there's an additional 'SetVisited' function in dirty flag that sets the
invalidation count to that one of the ref frame. This is exactly what previously only
happend when setting the input/output/op as invalite.