Gradient (a reference value type) can be a default
We needed this feature to hide default gradients in parameter groups. To accomplish this was tricky, because implementing the gradient editor in a way that returns cloned gradients if needed was tricky and sadly involved a lot of work arounds. Most of the changes are required, because in DrawEditControl() it wasn't possible to check if the value was a default value or not. So we also had to pass the input.