TransformImage2
Here is a proposal for a new version of the "TransformImage" tool. I removed the mirror setting and added a "TextureAddressMode" input parameter.
I duplicated the HLSL file and accordingly modified it.
I created an example symbol in order to describe the difference between both operators.
Descriptions have been updated as well.