Prevent sorting connectings on save because this can lead to inconsistencies in multi-inputs
The sort criteria only used targetNodeId and targetInputId. For multiInputs this leads to multiple connects with undefined sort order.

the order of objects with the same name is not guaranteed to remain intact when using .OrderBy(). The reason is that .OrderBy() is not a stable sort, meaning that it does not preserve the relative order of elements that are considered equal based on the sort criteria